refactor: extract streaming and output mapping helpers #37
